+++ src/Instrumentum_sanae_doctrinae/web_scraping/scrap_metadata.py
@@ -24,7 +24,7 @@ import asyncio
 import json 
 
 from ..web_scraping import http_connexion
-from ..web_scraping import _my_tools
+from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
 from ..web_scraping import my_constants
