+++ src/Instrumentum_sanae_doctrinae/web_scraping/sermonindex/si_scrap_metadata.py
@@ -16,7 +16,7 @@ from Instrumentum_sanae_doctrinae.web_scraping import http_connexion
 
 from Instrumentum_sanae_doctrinae.web_scraping import scrap_metadata
 from Instrumentum_sanae_doctrinae.web_scraping import my_constants
-from Instrumentum_sanae_doctrinae.web_scraping import _my_tools
+from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
