+++ src/Instrumentum_sanae_doctrinae/web_scraping/sermonindex/text_sermon/si_text_sermon_scrap_general_information.py
@@ -11,7 +11,9 @@ import json
 import urllib
 import urllib.parse
 
-from Instrumentum_sanae_doctrinae.web_scraping import _my_tools, http_connexion, my_constants
+from Instrumentum_sanae_doctrinae.web_scraping import http_connexion, my_constants
+from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
+
 from Instrumentum_sanae_doctrinae.web_scraping.sermonindex.si_scrap_metadata import SermonIndexScrapAuthorTopicScripturePage
