+++ src/Instrumentum_sanae_doctrinae/web_scraping/sermonindex/vintage_image/si_vin_im_scrap_work.py
@@ -6,11 +6,11 @@
 import json
 import os
 import pathlib
-import urllib
 
 import bs4 
-from bs4 import BeautifulSoup
-from Instrumentum_sanae_doctrinae.web_scraping import _my_tools, http_connexion, my_constants
+from Instrumentum_sanae_doctrinae.web_scraping import http_connexion, my_constants
+from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
+
 from Instrumentum_sanae_doctrinae.web_scraping.sermonindex.si_scrap_metadata import SermonIndexScrapAuthorTopicScripturePage
