+++ tests/monergism/test_mn_download.py
@@ -8,8 +8,8 @@ from Instrumentum_sanae_doctrinae.web_scraping.monergism import mn_download
 
 def test_download():
     root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
-    browse_by_type = "scripture"
-    name = "Genesis"
+    browse_by_type = "speaker"
+    name = "Archibald Alexander"
     ob = mn_download.MN_Download_Work(name,root_folder,browse_by_type)
     #print("Elvis")
     async def init_and_download():
